Reconfigured server into 2 files so test files can start a server
