+++ .circleci/config.yml
@@ -28,9 +28,9 @@ jobs:
       # If you are using yarn instead npm, remove the line above and uncomment the two lines below.
       # - node/install-packages:
       #     pkg-manager: yarn
-      - run:
-          name: Run server
-          command: npm run server-dev
+      # - run:
+      #     name: Run server
+      #     command: npm run server-dev
 
       - run:
           name: Run tests
