+++ tests/routes.test.js
@@ -1,8 +1,12 @@
-const request = require('supertest');
+const supertest = require('supertest');
+const app = require('../server/server');
+
+const request = supertest(app);
 
 describe('GET requests to the reviews endpoint', () => {
   it('respond with 200', (done) => {
-    request('http://localhost:3000')
+    request
+    // request('http://localhost:3000')
       .get('/reviews')
       .expect(200, done);
   });
@@ -10,7 +14,8 @@ describe('GET requests to the reviews endpoint', () => {
 
 describe('GET requests to the meta endpoint', () => {
   it('respond with 200', (done) => {
-    request('http://localhost:3000')
+    request
+    // request('http://localhost:3000')
       .get('/reviews/meta')
       .expect(200, done);
   });
@@ -18,7 +23,8 @@ describe('GET requests to the meta endpoint', () => {
 
 describe('POST requests to the reviews endpoint', () => {
   it('respond with 201', (done) => {
-    request('http://localhost:3000')
+    request
+    // request('http://localhost:3000')
       .post('/reviews')
       .expect(201, done);
   });
@@ -26,7 +32,8 @@ describe('POST requests to the reviews endpoint', () => {
 
 describe('PUT requests to the helpful endpoint', () => {
   it('respond with 204', (done) => {
-    request('http://localhost:3000')
+    request
+    // request('http://localhost:3000')
       .put('/reviews/16/helpful')
       .expect(204, done);
   });
@@ -34,7 +41,8 @@ describe('PUT requests to the helpful endpoint', () => {
 
 describe('PUT requests to the report endpoint', () => {
   it('respond with 204', (done) => {
-    request('http://localhost:3000')
+    request
+    // request('http://localhost:3000')
       .put('/reviews/16/report')
       .expect(204, done);
   });
